Fixed foreign key violation, added tests
Fixed foreig key constraints violations due to missing Mission ans Person.

modifies setEntities():
  -renamed to createTicket1AndInsert(): now the method creates only
    ticket1 and the related Mission and Person. and links them
if other Ticket are needed for tests, a method createTicket[num]AndInsert() should be created. This method should take care of the creation of the related Mission and Person, or it can use the same for testTicket1

Tests added:
-updateNotEmptyTicketTest() simple update test on a field
-updateTicketShouldNullEmptyFields() test the update on a ticket with some null fields.
-updateNullTicketIDShouldReturnZero()
-updateInvalidTicketIDShouldReturnZero()
-deleteExistingTicketTest()